* commit '16506d4ef2df8349ab0d6734111a53d46d0d910b':
  Change subId to int from long

+++ src/com/android/contacts/common/util/AccountSelectionUtil.java
@@ -55,12 +55,12 @@ public class AccountSelectionUtil {
 
         final private Context mContext;
         final private int mResId;
-        final private long mSubscriptionId;
+        final private int mSubscriptionId;
 
         final protected List<AccountWithDataSet> mAccountList;
 
         public AccountSelectedListener(Context context, List<AccountWithDataSet> accountList,
-                int resId, long subscriptionId) {
+                int resId, int subscriptionId) {
             if (accountList == null || accountList.size() == 0) {
                 Log.e(LOG_TAG, "The size of Account list is 0.");
             }
